automacao/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
psycopg2-binary>=2.9.0
google-generativeai>=0.8.0
boto3>=1.34.0
PyMuPDF>=1.23.0
langchain>=0.2.0
langchain-text-splitters>=0.2.0
tenacity>=8.2.0
